#69 refactoring package structure
