+++ crawler/src/main/java/org/zalando/apidiscovery/crawler/ClientsConfiguration.java
@@ -6,8 +6,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.StringHttpMessageConverter;
 import org.springframework.web.client.RestTemplate;
-import org.zalando.apidiscovery.crawler.storage.ApiDiscoveryStorageGateway;
-import org.zalando.apidiscovery.crawler.storage.LegacyApiDiscoveryStorageGateway;
+import org.zalando.apidiscovery.crawler.gateway.ApiDiscoveryStorageGateway;
+import org.zalando.apidiscovery.crawler.gateway.LegacyApiDiscoveryStorageGateway;
+import org.zalando.apidiscovery.crawler.gateway.WellKnownSchemaGateway;
 import org.zalando.stups.clients.kio.KioOperations;
 import org.zalando.stups.clients.kio.spring.RestTemplateKioOperations;
 import org.zalando.stups.oauth2.spring.client.StupsOAuth2RestTemplate;
