+++ crawler/src/test/java/org/zalando/apidiscovery/crawler/gateway/LegacyApiDiscoveryStorageGatewayTest.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.crawler.storage;
+package org.zalando.apidiscovery.crawler.gateway;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,12 +7,12 @@ import org.springframework.core.io.Resource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.zalando.apidiscovery.crawler.TestDataHelper.META_API_DEFINITION;
+import static org.zalando.apidiscovery.crawler.TestDataHelper.META_API_LEGACY_DEFINITION;
 import static org.zalando.apidiscovery.crawler.TestDataHelper.metaApiApplication;
 import static org.zalando.apidiscovery.crawler.TestDataHelper.readJson;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-public class ApiDiscoveryStorageGatewayTest {
+public class LegacyApiDiscoveryStorageGatewayTest {
 
     @Value("classpath:meta_api_schema_discovery.json")
     private Resource metaApiSchemaDiscoveryResource;
@@ -21,12 +21,12 @@ public class ApiDiscoveryStorageGatewayTest {
     private Resource metaApiDefinitionResource;
 
     @Test
-    public void shouldBeAbleToMapSchemaDiscoveryAndDefinitionToApiDefinition() throws Exception {
-        ApiDefinition apiDefinition = ApiDiscoveryStorageGateway.constructApiDefinition(
+    public void shouldBeAbleToMapSchemaDiscoveryAndDefinitionToLegacyApiDefinition() throws Exception {
+        LegacyApiDefinition legacyApiDefinition = LegacyApiDiscoveryStorageGateway.constructLegacyApiDefinition(
             readJson(metaApiSchemaDiscoveryResource),
             readJson(metaApiDefinitionResource),
             metaApiApplication());
 
-        assertThat(apiDefinition).isEqualTo(META_API_DEFINITION);
+        assertThat(legacyApiDefinition).isEqualTo(META_API_LEGACY_DEFINITION);
     }
 }
